src/inference/requirements.txt
catboost
uvicorn[standard]
fastapi
scikit-learn
python-dotenv
pydantic-settings